refactor(agent_memory): move memory_loader into agent_memory domain
Move the canonical `memory_loader.rs` (MemoryLoader trait,
DefaultMemoryLoader, MemoryCitation, cross-chat context, prior
conversations) from `agent/memory_loader.rs` into the new
`agent_memory/memory_loader.rs` domain module.

The original file becomes a thin re-export shim so all existing
harness imports (`agent::memory_loader::*`) continue to resolve.
