+++ src/openhuman/agent_memory/mod.rs
@@ -6,5 +6,6 @@
 //! tree browsing to answer queries. This domain centralizes the agent definition,
 //! prompt construction, and retrieval performance tracking.
 
+pub mod memory_loader;
 pub mod ops;
 pub mod types;
